requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
redis>=4.5.0
pydantic>=1.8.0
numpy>=1.21.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.52
langchain-openai>=0.0.5
langchain-redis>=0.0.5
PyPDF2>=3.0.0
tiktoken>=0.5.2
pydantic-settings>=2.0.0
pypdf>=3.11.0
